safe newlabel deleted
